add TitleScene
