+++ src/Scenes/PreloaderScene.js
@@ -1,5 +1,4 @@
 import Phaser from "phaser";
-import ScrollingBackground from "../Entities/Scrolling";
 export default class PreloaderScene extends Phaser.Scene {
   constructor() {
     super({ key: "PreloaderScene" });
@@ -43,6 +42,7 @@ export default class PreloaderScene extends Phaser.Scene {
     // update progress bar
     this.load.on('progress', (value) => {
       percentText.setText(parseInt(value * 100) + '%');
+      
       progressBar.clear();
       progressBar.fillStyle(0xffffff, 1);
       progressBar.fillRect(250, 280, 300 * value, 30);
@@ -50,14 +50,9 @@ export default class PreloaderScene extends Phaser.Scene {
 
 
     // remove progress bar when complete
-    this.load.on('complete', function () {
-      progressBar.destroy();
-      progressBox.destroy();
-      loadingText.destroy();
-      percentText.destroy();
-     
+    this.load.on('complete', () => {
       this.ready();
-    }.bind(this));
+    });
 
     this.timedEvent = this.time.delayedCall(3000, this.ready, [], this);
  
@@ -75,67 +70,14 @@ export default class PreloaderScene extends Phaser.Scene {
     this.load.audio("sndBtnDown", "assets/audio/sndBtnDown.wav");
   }
   create() {
-    this.sfx = {
-      btnOver: this.sound.add("sndBtnOver"),
-      btnDown: this.sound.add("sndBtnDown")
-    };
-    // this.scene.start("SceneMain");
-    this.btnPlay = this.add.sprite(
-      this.game.config.width * 0.5,
-      this.game.config.height * 0.5,
-      "sprBtnPlay"
-      
-    );
-    this.btnPlay.setInteractive();
-    this.btnPlay.on("pointerover", () => {
-    this.btnPlay.setTexture("sprBtnPlayHover"); // set the button texture to sprBtnPlayHover
-    this.sfx.btnOver.play(); // play the button over sound
-    }, this);
-    // this.btnPlay.on("pointerout", () => {
-    //   this.setTexture("sprBtnPlay");
-    // });
-
-    this.btnPlay.on("pointerdown",() =>{
-      this.btnPlay.setTexture("sprBtnPlayDown");
-      this.sfx.btnDown.play();
-    }, this)
-
-    // this.btnPlay.on("pointerup", () => {
-    //   this.setTexture("sprBtnPlay");
-    // }, this);
-
-    this.btnPlay.on("pointerup", () => {
-      this.btnPlay.setTexture("sprBtnPlay");
-      this.scene.start("SceneMain");
-    }, this);
-
-    this.title = this.add.text(this.game.config.width * 0.5, 128, "SPACE SHOOTER", {
-      fontFamily: 'monospace',
-      fontSize: 48,
-      fontStyle: 'bold',
-      color: '#fffffa',
-      align: 'center'
-    });
-    this.title.setOrigin(0.5);
-
-    this.backgrounds = [];
-for (let i = 0; i < 5; i++) {
-  let keys = ["sprBg0", "sprBg1"];
-  let key = keys[Phaser.Math.Between(0, keys.length - 1)];
-  let bg = new ScrollingBackground(this, key, i * 10);
-  this.backgrounds.push(bg);
-}
+    
   }
-update(){
-  for (let i = 0; i < this.backgrounds.length; i++) {
-    this.backgrounds[i].update();
-  }
-}
+
 
 ready() {
   this.readyCount += 1;
   if (this.readyCount === 2) {
-    this.scene.start('Title');
+    this.scene.start('TitleScene');
   }
 }
 }
